requirements.txt
kaggle
python-dotenv